ModelAPI/requirements.txt
fastapi
uvicorn
torch
ultralytics
pillow
